C# plan docs/structure review
