+++ boltffi_bindgen/src/render/csharp/plan/identifier.rs
@@ -1,22 +1,14 @@
-//! Identifiers for C-side FFI symbols. The C# identifiers (class /
-//! method / property / param / local / namespace names) live in
-//! [`super::super::ast::identifier`]: they speak the C# grammar and
-//! belong to the AST layer. This file holds the one identifier that
-//! names a *C* symbol crossing the ABI boundary.
-//!
-//! Splitting `C*` from `CSharp*` at the type-name level (and now at
-//! the module level) tells the reader at a glance which side of the
-//! boundary each value sits on.
-
 use std::fmt;
 
 use boltffi_ffi_rules::naming::{GlobalSymbol, Name};
 
-/// The name of a C function exported from the native library
-/// (e.g., `"boltffi_echo_i32"`, `"boltffi_free_buf"`). Goes inside a
-/// `DllImport` `EntryPoint` attribute. The lowerer constructs the
-/// complete symbol using the `naming::*` helpers and wraps it with
-/// [`Self::new`].
+/// The C-side function symbol used as the `EntryPoint` of a `[DllImport]` attribute.
+///
+/// Examples:
+/// ```csharp
+/// [DllImport(LibName, EntryPoint = "boltffi_echo_i32")]
+/// //                                ^^^^^^^^^^^^^^^^
+/// ```
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
 pub struct CFunctionName(String);
