+++ boltffi_bindgen/src/render/csharp/plan/mod.rs
@@ -1,11 +1,13 @@
-//! View model for the C# backend: the FFI-shaped data the templates
-//! consume. Records, enums, functions, methods, params: each plan type
-//! models an FFI shape and carries
-//! [`super::ast`](super::ast) values as field payloads.
-//! `plan` speaks FFI vocabulary; `ast` speaks C# grammar.
+//! Plans are the primary data structure used by the `templates` module to produce C# code. Thus,
+//! they are the `ViewModel` of the `templates` `View`.
 //!
-//! No dependency on `emit` or `lower`: the plan is passive data.
-//! `lower` produces it, `templates` consume it.
+//! Plans consist of `ast` (C# grammar) primitive values, and are structured around the FFI boundary
+//! code being generated.
+//!
+//! Templates contain most of the complexity of the C# syntax, but plans and their `ast` primitives
+//! exist only to expose conditional or variable syntax.
+//!
+//! Plans are created by the `lower` module, and consumed by the `templates` module.
 
 mod callable;
 mod enumeration;
@@ -15,10 +17,10 @@ mod module;
 mod record;
 
 pub use callable::{
-    CSharpFunctionPlan, CSharpMethodPlan, CSharpParamPlan, CSharpParamKind, CSharpReceiver, CSharpReturnKind,
-    CSharpWireWriterPlan,
+    CSharpFunctionPlan, CSharpMethodPlan, CSharpParamKind, CSharpParamPlan, CSharpReceiver,
+    CSharpReturnKind, CSharpWireWriterPlan,
 };
-pub use enumeration::{CSharpEnumPlan, CSharpEnumKind, CSharpEnumVariantPlan};
+pub use enumeration::{CSharpEnumKind, CSharpEnumPlan, CSharpEnumVariantPlan};
 pub use field::CSharpFieldPlan;
 pub use identifier::CFunctionName;
 pub use module::CSharpModulePlan;
